fix: re-arrange test cases for backend
