+++ backend/src/contracts/mod.rs
@@ -1,4 +1,3 @@
 pub mod generated;
 pub mod mock;
 pub(crate) mod signer;
-pub(crate) mod tests;
