+++ backend/src/lib.rs
@@ -1,3 +1,4 @@
 #![feature(generic_const_exprs)]
 pub mod apis;
 pub mod contracts;
+pub mod tests;
